make sure lapse % doesn't create decimal intervals
issue reported on
https://github.com/dae/anki/pull/272

+++ anki/schedv2.py
@@ -859,7 +859,7 @@ select id from cards where did in %s and queue = 2 and due <= ? limit ?)"""
         return delay
 
     def _lapseIvl(self, card, conf):
-        ivl = max(1, conf['minInt'], card.ivl*conf['mult'])
+        ivl = max(1, conf['minInt'], int(card.ivl*conf['mult']))
         return ivl
 
     def _rescheduleRev(self, card, ease, early):
